refactor : 승무패 enum Domain용과 View용으로 분리
