+++ src/main/java/blackjack/domain/WinningResult.java
@@ -1,0 +1,12 @@
+package blackjack.domain;
+
+import java.util.List;
+
+public enum WinningResult {
+
+    WIN,
+    LOSE,
+    PUSH,
+    ;
+
+}
