+++ src/main/java/blackjack/view/OutputView.java
@@ -3,8 +3,7 @@ package blackjack.view;
 import blackjack.domain.Participant;
 import blackjack.domain.Participants;
 import blackjack.domain.Player;
-import blackjack.domain.CardSuit;
-import blackjack.util.WinningResult;
+import blackjack.domain.WinningResult;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -62,14 +61,14 @@ public class OutputView {
 
     public void printDealerWinORLose(List<WinningResult> dealerResult) {
         System.out.print("딜러: ");
-        System.out.print(WinningResult.WIN.winCount(dealerResult) + WinningResult.WIN.getName());
-        System.out.print(WinningResult.PUSH.pushCount(dealerResult) + WinningResult.PUSH.getName());
-        System.out.println(WinningResult.LOSE.loseCount(dealerResult) + WinningResult.LOSE.getName());
+        System.out.print(ViewWinningResult.WIN.winCount(dealerResult) + ViewWinningResult.WIN.getName());
+        System.out.print(ViewWinningResult.PUSH.pushCount(dealerResult) + ViewWinningResult.PUSH.getName());
+        System.out.println(ViewWinningResult.LOSE.loseCount(dealerResult) + ViewWinningResult.LOSE.getName());
     }
 
     public void printPlayerWinORLose(Map<Player, WinningResult> playerResult) {
         for (Participant player : playerResult.keySet()) {
-            System.out.println(player.getName() + " : " + playerResult.get(player).getName());
+            System.out.println(player.getName() + " : " + ViewWinningResult.findWinningResult(playerResult.get(player)).getName());
         }
     }
 
@@ -89,7 +88,7 @@ public class OutputView {
             ViewCardSuit cardSuit = ViewCardSuit.findCardSuit(participant.getCardSuit(i));
             card.add(cardNumber.getName() + cardSuit.getCardSuitName());
         }
-        System.out.printf(String.join(", ",card));
+        System.out.printf(String.join(", ", card));
         System.out.println();
     }
 }
