+++ src/main/java/blackjack/view/ViewWinningResult.java
@@ -1,17 +1,21 @@
-package blackjack.util;
+package blackjack.view;
 
+import blackjack.domain.WinningResult;
+import java.util.Arrays;
 import java.util.List;
 
-public enum WinningResult {
+public enum ViewWinningResult {
 
-    WIN("승"),
-    LOSE("패"),
-    PUSH("무"),
+    WIN(WinningResult.WIN, "승"),
+    LOSE(WinningResult.LOSE, "패"),
+    PUSH(WinningResult.PUSH, "무"),
     ;
 
+    final WinningResult winningResult;
     final String name;
 
-    WinningResult(String name) {
+    ViewWinningResult(WinningResult winningResult, String name) {
+        this.winningResult = winningResult;
         this.name = name;
     }
 
@@ -31,6 +35,13 @@ public enum WinningResult {
         return (int) dealerResult.stream()
             .filter(result -> result.equals(PUSH))
             .count();
+    }
+
+    public static ViewWinningResult findWinningResult(WinningResult winningResult) {
+        return Arrays.stream(ViewWinningResult.values())
+            .filter(viewWinningResult -> viewWinningResult.winningResult == winningResult)
+            .findFirst()
+            .get();
     }
 
     public String getName() {
